+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 
 ### Enhancements:
 
-- feat(tls_custom_certificate): Add support for allow_untrusted_root attribute
+- feat(ngwaf/v1/workspaces/thresholds): adds CRUD support for NGWAF Thresholds ([#713](https://github.com/fastly/go-fastly/pull/713))
+- feat(tls_custom_certificate): Add support for allow_untrusted_root attribute ([#596](https://github.com/fastly/go-fastly/pull/596))
 
 ### Bug fixes: